ZQL Refactor part 2
