+++ src/main/java/edu/uiuc/zenvisage/zql/filters/ArgMinSortFilterPrimitive.java
@@ -1,10 +1,12 @@
 /**
  * 
  */
-package edu.uiuc.zenvisage.zqlcomplete.querygraph;
+package edu.uiuc.zenvisage.zql.filters;
 
 import java.util.ArrayList;
 import java.util.List;
+
+import edu.uiuc.zenvisage.zql.AxisVariableScores;
 
 /**
  * @author tarique
