+++ src/main/java/edu/uiuc/zenvisage/zql/functions/D.java
@@ -1,12 +1,14 @@
 /**
  * @author Tarique
  */
-package edu.uiuc.zenvisage.zqlcomplete.querygraph;
+package edu.uiuc.zenvisage.zql.functions;
 
 import java.util.List;
 
 import edu.uiuc.zenvisage.data.remotedb.VisualComponent;
 import edu.uiuc.zenvisage.data.remotedb.VisualComponentList;
+import edu.uiuc.zenvisage.zql.AxisVariable;
+import edu.uiuc.zenvisage.zql.AxisVariableScores;
 
 /**
  * @author tarique
