+++ src/main/java/edu/uiuc/zenvisage/zql/package-info.java
@@ -6,4 +6,4 @@
  * @author eideh
  *
  */
-package edu.uiuc.zenvisage.zqlcomplete.querygraph;
+package edu.uiuc.zenvisage.zql;
